# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Business Validation Platform Document API
  version: 0.12.0-phase3-taskrun
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: >
    Canonical authentication, document upload, asynchronous processing,
    structured-plan mutation, and explicit confirmation contract.
servers:
  - url: /api/v1
...
security:
  - platformAuth: []
paths:
  /auth/signup:
    post:
      tags: [Authentication]
      summary: Register an active user and issue an access/refresh pair
      operationId: signup
      x-implementation-phase: Phase 3
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SignupRequest'}
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthApiResponse'}
        '400': {$ref: '#/components/responses/StandardError'}
        '409': {$ref: '#/components/responses/StandardError'}
  /auth/login:
    post:
      tags: [Authentication]
      summary: Authenticate with normalized email and password
      operationId: login
      x-implementation-phase: Phase 3
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '200':
          description: Authentication succeeded
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthApiResponse'}
        '400': {$ref: '#/components/responses/StandardError'}
        '401': {$ref: '#/components/responses/StandardError'}
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Rotate a valid refresh token and issue a new pair
      operationId: refreshAccessToken
      x-implementation-phase: Phase 3
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RefreshRequest'}
      responses:
        '200':
          description: Token pair rotated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TokenPairApiResponse'}
        '400': {$ref: '#/components/responses/StandardError'}
        '401': {$ref: '#/components/responses/StandardError'}
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Revoke the supplied current refresh token
      operationId: logout
      x-implementation-phase: Phase 3
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RefreshRequest'}
      responses:
        '200':
          description: Refresh token revoked
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EmptyApiResponse'}
        '400': {$ref: '#/components/responses/StandardError'}
        '401': {$ref: '#/components/responses/StandardError'}
  /users/me:
    get:
      tags: [Users]
      summary: Get the current authenticated user
      operationId: getCurrentUser
      x-implementation-phase: Phase 3
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UserApiResponse'}
        '401': {$ref: '#/components/responses/StandardError'}
...
  /projects/{projectId}/documents:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      tags: [Documents]
      summary: Upload a business plan and queue document processing
      operationId: uploadProjectDocument
      x-implementation-phase: Phase 1C
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: DOCX only, maximum 20 MiB
                documentType:
                  type: string
                  enum: [BUSINESS_PLAN]
                  default: BUSINESS_PLAN
      responses:
        '202':
          description: Upload stored and processing queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentUploadApiResponse'
        '400':
          $ref: '#/components/responses/StandardError'
        '404':
          $ref: '#/components/responses/StandardError'
        '409':
          $ref: '#/components/responses/StandardError'
        '413':
          $ref: '#/components/responses/StandardError'
        '415':
          $ref: '#/components/responses/StandardError'
        '422':
          $ref: '#/components/responses/StandardError'
        '503':
          $ref: '#/components/responses/StandardError'
    get:
      tags: [Documents]
      summary: List project documents
      operationId: listProjectDocuments
      x-implementation-phase: Phase 1B
      responses:
        '200':
          description: Document list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListApiResponse'
        '404':
          $ref: '#/components/responses/StandardError'
  /documents/{documentId}/versions/{versionId}:
    get:
      tags: [Documents]
      summary: Get one document version
      operationId: getDocumentVersion
      x-implementation-phase: Phase 1B
      parameters:
        - $ref: '#/components/parameters/DocumentId'
        - $ref: '#/components/parameters/VersionId'
      responses:
        '200':
          description: Document version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentVersionApiResponse'
        '404':
          $ref: '#/components/responses/StandardError'
...
  /projects/{projectId}/financial-analyses:
    get:
      tags: [Financial Analyses]
      summary: List non-deleted project financial analyses
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200': {description: Financial analysis summaries}
    post:
      tags: [Financial Analyses]
      summary: Create a financial-analysis draft from confirmed assumptions
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '201': {description: Financial analysis draft}
  /projects/{projectId}/financial-analyses/{analysisId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/AnalysisId'
    get:
      tags: [Financial Analyses]
      summary: Get a financial analysis and its immutable source snapshot
      responses:
        '200': {description: Financial analysis detail}
    patch:
      tags: [Financial Analyses]
      summary: Update a draft financial analysis
      responses:
        '200': {description: Updated financial analysis}
    delete:
      tags: [Financial Analyses]
      summary: Soft-delete a financial analysis
      responses:
        '204': {description: Deleted}
  /projects/{projectId}/financial-analyses/{analysisId}/run:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/AnalysisId'
    post:
      tags: [Financial Analyses]
      summary: Deterministically calculate and complete a financial analysis
      responses:
        '200': {description: Completed financial analysis}
  /projects/{projectId}/financial-analyses/{analysisId}/duplicate:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/AnalysisId'
    post:
      tags: [Financial Analyses]
      summary: Duplicate assumptions into a new financial-analysis draft
      responses:
        '201': {description: New financial analysis draft}
...
components:
  securitySchemes:
    platformAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Phase 3 access JWT. The development X-User-Id adapter is profile-bound
        and is deliberately excluded from the public contract.
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: {type: integer, format: int64}
...
    AnalysisId:
      name: analysisId
      in: path
      required: true
      schema: {type: integer, format: int64}
    DocumentId:
      name: documentId
      in: path
      required: true
      schema: {type: integer, format: int64}
    VersionId:
      name: versionId
      in: path
      required: true
      schema: {type: integer, format: int64}
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: >
        Client-generated retry key. After trimming, 1-100 ASCII characters
        from A-Z, a-z, 0-9, period, underscore, colon, and hyphen are allowed.
      schema:
        type: string
        minLength: 1
        maxLength: 100
        pattern: '^[A-Za-z0-9._:-]+$'
...
  responses:
...
    StandardError:
      description: Standard API error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorApiResponse'
  schemas:
...
    Meta:
      type: object
      required: [requestId, timestamp]
      properties:
        requestId: {type: string}
        timestamp: {type: string, format: date-time}
    ApiError:
      type: object
      required: [code, message, fieldErrors, retryable]
      properties:
        code: {type: string}
        message: {type: string}
        fieldErrors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: {type: string}
              message: {type: string}
        retryable: {type: boolean}
    ErrorApiResponse:
      type: object
      required: [success, error, meta]
      properties:
        success: {type: boolean, const: false}
        error: {$ref: '#/components/schemas/ApiError'}
        meta: {$ref: '#/components/schemas/Meta'}
    SignupRequest:
      type: object
      required: [email, password, displayName]
      properties:
        email: {type: string, format: email, maxLength: 254}
        password:
          type: string
          format: password
          minLength: 8
          description: Maximum 72 UTF-8 bytes; never persisted or logged raw.
        displayName: {type: string, minLength: 1, maxLength: 50}
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email, maxLength: 254}
        password: {type: string, format: password}
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          minLength: 1
          description: Raw token is accepted only at this boundary and stored as SHA-256.
    User:
      type: object
      required: [id, email, displayName, role, status]
      properties:
        id: {type: integer, format: int64}
        email: {type: string, format: email}
        displayName: {type: string}
        role: {type: string, enum: [USER, ADMIN]}
        status:
          type: string
          enum: [PENDING, ACTIVE, LOCKED, SUSPENDED, WITHDRAWN]
    TokenPair:
      type: object
      required:
        [tokenType, accessToken, accessTokenExpiresIn, refreshToken,
         refreshTokenExpiresIn]
      properties:
        tokenType: {type: string, const: Bearer}
        accessToken: {type: string}
        accessTokenExpiresIn: {type: integer, format: int64, minimum: 1}
        refreshToken: {type: string}
        refreshTokenExpiresIn: {type: integer, format: int64, minimum: 1}
    Auth:
      type: object
      required: [user, tokens]
      properties:
        user: {$ref: '#/components/schemas/User'}
        tokens: {$ref: '#/components/schemas/TokenPair'}
...
    DocumentUpload:
      type: object
      required: [projectId, documentId, versionId, jobId, status]
      properties:
        projectId: {type: integer, format: int64}
        documentId: {type: integer, format: int64}
        versionId: {type: integer, format: int64}
        jobId: {type: integer, format: int64}
        status:
          $ref: '#/components/schemas/JobStatus'
          description: QUEUED for a newly accepted upload; current job status for an idempotent replay.
    DocumentSummary:
      type: object
      required: [documentId, documentType, currentVersion, status, latestVersionId, updatedAt]
      properties:
        documentId: {type: integer, format: int64}
        documentType: {type: string, enum: [BUSINESS_PLAN]}
        currentVersion: {type: integer}
        status: {type: string, enum: [ACTIVE, REPLACED, DELETED]}
        latestVersionId: {type: integer, format: int64}
        updatedAt: {type: string, format: date-time}
    DocumentVersion:
      type: object
      required: [documentId, versionId, versionNumber, originalFileName, sizeBytes, parseStatus, uploadedAt]
      properties:
        documentId: {type: integer, format: int64}
        versionId: {type: integer, format: int64}
        versionNumber: {type: integer}
        originalFileName: {type: string}
        sizeBytes: {type: integer, format: int64}
        parseStatus: {$ref: '#/components/schemas/JobStatus'}
        uploadedAt: {type: string, format: date-time}
    JobStatus:
      type: string
      enum:
        - NOT_STARTED
        - QUEUED
        - RUNNING
        - SUCCEEDED
        - PARTIAL
        - FAILED
        - CANCELED
...
    DocumentUploadApiResponse:
      type: object
      required: [success, data, meta]
      properties:
        success: {type: boolean, const: true}
        data: {$ref: '#/components/schemas/DocumentUpload'}
        meta: {$ref: '#/components/schemas/Meta'}
    DocumentListApiResponse:
      type: object
      required: [success, data, meta]
      properties:
        success: {type: boolean, const: true}
        data:
          type: array
          items: {$ref: '#/components/schemas/DocumentSummary'}
        meta: {$ref: '#/components/schemas/Meta'}
    DocumentVersionApiResponse:
      type: object
      required: [success, data, meta]
      properties:
        success: {type: boolean, const: true}
        data: {$ref: '#/components/schemas/DocumentVersion'}
        meta: {$ref: '#/components/schemas/Meta'}
...
    AuthApiResponse:
      type: object
      required: [success, data, meta]
      properties:
        success: {type: boolean, const: true}
        data: {$ref: '#/components/schemas/Auth'}
        meta: {$ref: '#/components/schemas/Meta'}
...
    TokenPairApiResponse:
      type: object
      required: [success, data, meta]
      properties:
        success: {type: boolean, const: true}
        data: {$ref: '#/components/schemas/TokenPair'}
        meta: {$ref: '#/components/schemas/Meta'}
    UserApiResponse:
      type: object
      required: [success, data, meta]
      properties:
        success: {type: boolean, const: true}
        data: {$ref: '#/components/schemas/User'}
        meta: {$ref: '#/components/schemas/Meta'}
    EmptyApiResponse:
      type: object
      required: [success, meta]
      properties:
        success: {type: boolean, const: true}
        meta: {$ref: '#/components/schemas/Meta'}
...